Collapse to self-contained install.py + scripts/core.py
Per revised PRD (#1) and revised acceptance criteria on #3:

- Drop scripts/lib/ package; collapse helpers into one scripts/core.py
- Drop scripts/list.py and scripts/paths.py — replaced by inline Bash in SKILL.md
- Drop the corresponding lib unit tests; install.py exercises core.py end-to-end
- Fix direct-invocation bug: scripts/install.py now runs as
  `python3 scripts/install.py ...` without ImportError (no relative imports,
  no package context required)

3 tests pass; direct --help works.
