+++ tests/conftest.py
@@ -1,4 +1,3 @@
-import os
 import subprocess
 import sys
 from pathlib import Path
@@ -6,7 +5,7 @@ from pathlib import Path
 import pytest
 
 ROOT = Path(__file__).resolve().parent.parent
-sys.path.insert(0, str(ROOT))
+sys.path.insert(0, str(ROOT / "scripts"))
 
 
 @pytest.fixture
@@ -22,14 +21,11 @@ def home(tmp_path, monkeypatch):
 def fake_upstream_repo(tmp_path):
     """Factory: build a local git repo with given files at given subpath.
 
-    Usage:
-        repo_url = fake_upstream_repo("acme/widget", "skills/widget", {
-            "SKILL.md": "---\\nname: widget\\n---\\n# v1\\n",
-        })
+    Returns a file:// URL git can clone from.
     """
     counter = {"n": 0}
 
-    def _make(repo_id: str, subpath: str, files: dict[str, str], extra_commits: list[dict[str, str]] | None = None) -> str:
+    def _make(repo_id: str, subpath: str, files: dict[str, str]) -> str:
         counter["n"] += 1
         repo_dir = tmp_path / f"upstream-{counter['n']}"
         repo_dir.mkdir()
@@ -37,7 +33,7 @@ def fake_upstream_repo(tmp_path):
         run("git", "init", "-q", "-b", "main")
         run("git", "config", "user.email", "test@test")
         run("git", "config", "user.name", "test")
-        target = repo_dir / subpath
+        target = repo_dir / subpath if subpath and subpath != "." else repo_dir
         target.mkdir(parents=True, exist_ok=True)
         for rel, content in files.items():
             p = target / rel
@@ -45,11 +41,6 @@ def fake_upstream_repo(tmp_path):
             p.write_text(content)
         run("git", "add", "-A")
         run("git", "commit", "-q", "-m", "initial")
-        for commit in extra_commits or []:
-            for rel, content in commit.items():
-                (target / rel).write_text(content)
-            run("git", "add", "-A")
-            run("git", "commit", "-q", "-m", "update")
         return f"file://{repo_dir}"
 
     return _make
